// frontend/src/external/client/api/orval/generated/models/modelsCreateEducationRequest.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * inselfy API
 * OpenAPI spec version: 1.0.0
 */

/**
 * 学歴作成リクエスト
 */
export interface ModelsCreateEducationRequest {
  /**
     * 学校名
     * @minLength 1
     * @maxLength 200
     */
  school: string;
  /**
     * 学部・学科・学位
     * @maxLength 200
     * @nullable
     */
  degree?: string | null;
  /**
     * 入学年
     * @minimum 1950
     * @maximum 2100
     * @nullable
     */
  startYear?: number | null;
  /**
     * 卒業年
     * @minimum 1950
     * @maximum 2100
     * @nullable
     */
  endYear?: number | null;
}